Some refactor to take out duplicate code
