+++ cs/src/devices/AzureStorageDevice/AzureStorageCommitManager.cs
@@ -45,23 +45,7 @@ namespace FASTER.devices
         // TODO(Tianyu): This is duplicate from CheckpointManager, should reuse when possible
         public byte[] GetCommitMetadata()
         {
-            // Assuming that the page blob already exists
-            int length = ReadInt32(blob, 0);
-            byte[] result = new byte[length];
-            var downloaded = blob.DownloadRangeToByteArray(result, 0, sizeof(Int32), length);
-            Debug.Assert(downloaded == length, "Underfilled read buffer");
-            return result;
+            return BlobUtil.ReadMetadataFile(blob);
         }
-        
-        private int ReadInt32(CloudPageBlob blob, long offset)
-        {
-            byte[] result = new byte[sizeof(Int32)];
-            var read = blob.DownloadRangeToByteArray(result, 0, offset, sizeof(Int32));
-            // TODO(Tianyu): Can read bytes ever be smaller than requested like POSIX? There is certainly
-            // no documentation about the behavior...
-            Debug.Assert(read == sizeof(Int32), "Underfilled read buffer");
-            return BitConverter.ToInt32(result, 0);
-        }
-
     }
 }
